Fix local dependencies issues (#5)
## Overview
Fixes build issues introduced by the local dependency `gedinline.jar` by installing the JAR into the local repository in the container during the build process. Updates documentation for building the JAR file and Github packages workflows to reflect this new dependency. Dockerfile is also updated to access Valipop jar file corresponding to the version number in `pom.xml`, rather than having to change the Dockerfile for every incremental release.

## Commits included
* Updated Docs for local dependencies
     * Updated README and docs to account for the addition of local
dependencies which must be installed before building jar.
* Updated .gitignore to ignore .DS_Store
     * Updated .gitignore to ignore MacOS .DS_Store files and removed an
instance of this which had been committed to the repository.
* Fixes docker build issues
     * Fixes docker build issues with local dependencies.
* Updated Dockerfile to use latest version
     * Updates dockerfile to use latest version from the pom.xml, rather than a
hardcoded path to a JAR file which needs to be updated for every new
release.
* Updates version number
     * Updates version number in pom.xml from 4.0-SNAPSHOT to 4.1-SNAPSHOT to
line up with latest version release.
* Updates build-and-test workflow
     * Updates build-and-test workflow to install local dependency.